feat(frontend): add profile store and refactor
